+++ libs/widgets-shared/index.ts
@@ -28,3 +28,11 @@ export function getResourceText(text: ResourceText, locale: string): string {
   }
   return text[locale] || text['en'] || 'Unknown';
 }
+
+// Difference between Windows epoch (1601) and Unix epoch (1970) in milliseconds
+const EPOCH_DIFF_MILLISECONDS = 11644473600000n;
+
+/** Convert Windows FileTime to Js Unix Date */
+export function WindowsDateFileTimeToDate(fileTime: bigint) {
+  return new Date(Number(fileTime / 10000n - EPOCH_DIFF_MILLISECONDS));
+}