+++ changelog.md
@@ -5,6 +5,7 @@
 - third party widgets sometimes not starting.
 - can not save settings on empty wallpaper list.
 - not updating UI on remotion of resources.
+- button to update downloaded resources not working properly.
 
 ## [2.4.1]
 ### features